fix: Don't subscribe to Store change, but save it before closing the app
